requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
xgboost==2.0.3
scikit-learn==1.4.2
psycopg2-binary==2.9.9
schedule==1.2.2
